fix: extract ProcessVariableValue into standalone class
